requirements.txt
# pip3 install -r requirements.txt to install these dependencies
flask
python-dotenv
python-jose
flask-cors
six
pymongo
dnspython
flask_limiter
flasgger